chore(testing): refactor e2e-workspace tests to e2e-nx tests
remove run commands

remove workspace core e2e

chore(testing): remove e2e-cli package

chore(testing): add nx-run e2e package

chore(testing): refactor cache tests

chore(testing): remove empty workspace-core e2e lib

chore(testing): update graph asset
