+++ e2e/nx-misc/tsconfig.json
