+++ e2e/nx-run/tsconfig.spec.json
