--- e2e/workspace-core/jest.config.ts
@@ -1,10 +1,0 @@
-export default {
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-  maxWorkers: 1,
-  globals: { 'ts-jest': { tsconfig: '<rootDir>/tsconfig.spec.json' } },
-  displayName: 'e2e-workspace-core',
-  preset: '../../jest.preset.js',
-};
