+++ workspace.json
@@ -16,7 +16,6 @@
     "e2e-add-nx-to-monorepo": "e2e/add-nx-to-monorepo",
     "e2e-angular-core": "e2e/angular-core",
     "e2e-angular-extensions": "e2e/angular-extensions",
-    "e2e-cli": "e2e/cli",
     "e2e-cypress": "e2e/cypress",
     "e2e-detox": "e2e/detox",
     "e2e-jest": "e2e/jest",
@@ -25,15 +24,15 @@
     "e2e-next": "e2e/next",
     "e2e-node": "e2e/node",
     "e2e-nx-init": "e2e/nx-init",
+    "e2e-nx-misc": "e2e/nx-misc",
+    "e2e-nx-run": "e2e/nx-run",
     "e2e-nx-plugin": "e2e/nx-plugin",
     "e2e-react": "e2e/react",
     "e2e-react-native": "e2e/react-native",
     "e2e-storybook": "e2e/storybook",
     "e2e-utils": "e2e/utils",
     "e2e-web": "e2e/web",
-    "e2e-workspace-core": "e2e/workspace-core",
     "e2e-workspace-create": "e2e/workspace-create",
-    "e2e-workspace-integrations": "e2e/workspace-integrations",
     "eslint-plugin-nx": "packages/eslint-plugin-nx",
     "eslint-rules": "tools/eslint-rules",
     "express": "packages/express",
